added Spring Security

+++ src/main/java/com/aman/project/airBnbApp/service/RoomServiceImpl.java
@@ -3,7 +3,9 @@ package com.aman.project.airBnbApp.service;
 import com.aman.project.airBnbApp.dto.RoomDto;
 import com.aman.project.airBnbApp.entity.Hotel;
 import com.aman.project.airBnbApp.entity.Room;
+import com.aman.project.airBnbApp.entity.User;
 import com.aman.project.airBnbApp.exception.ResourceNotFoundException;
+import com.aman.project.airBnbApp.exception.UnAuthorisedException;
 import com.aman.project.airBnbApp.repository.HotelRepository;
 import com.aman.project.airBnbApp.repository.RoomRepository;
 import jakarta.transaction.Transactional;
@@ -12,6 +14,7 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -31,6 +34,11 @@ public class RoomServiceImpl implements RoomService {
 			.findById(hotelId)
 			.orElseThrow(() -> new ResourceNotFoundException("Hotel not found with Id " + hotelId));
 
+		User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		if (!user.equals(hotel.getOwner())) {
+			throw new UnAuthorisedException("This user does not own this hotel with id: " + hotelId);
+		}
+
 		Room room = modelMapper.map(roomDto, Room.class);
 		room.setHotel(hotel);
 		room = roomRepository.save(room);
@@ -47,6 +55,11 @@ public class RoomServiceImpl implements RoomService {
 		Hotel hotel = hotelRepository
 			.findById(hotelId)
 			.orElseThrow(() -> new ResourceNotFoundException("Hotel not found with Id " + hotelId));
+
+		//		User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		//		if (!user.equals(hotel.getOwner())) {
+		//			throw new UnAuthorisedException("This user does not own this hotel with id: " + hotelId);
+		//		}
 
 		return hotel
 			.getRooms()
@@ -72,7 +85,12 @@ public class RoomServiceImpl implements RoomService {
 		Room room = roomRepository
 			.findById(roomId)
 			.orElseThrow(() -> new ResourceNotFoundException("Room not found with Id " + roomId));
-		// TODO: some error , cant delete the rooms
+
+		User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		if (!user.equals(room.getHotel().getOwner())) {
+			throw new UnAuthorisedException("This user does not own this room with id: " + roomId);
+		}
+
 		inventoryService.deleteAllInventories(room);
 		roomRepository.deleteById(roomId);
 	}
